- Simple order by member - Checking order today (all order by admin)

+++ src/main/java/com/andy/sapofnbcrawler/response/MenuResponse.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonRootName;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -17,8 +18,8 @@ public class MenuResponse {
     
     @Data
     public static class DishResponse {
-        private String name;
-        private int price;
+        private String     name;
+        private BigDecimal price;
     }
     
 }
